split blockchain test group into two parallel CI jobs
The blockchain test group takes ~60 min because test_blockchain.py
contains 87% of all test instances (1362/1566). Split into two
subdirectories so build-job-matrix.py creates separate CI jobs:

- blockchain.validation (~422 tests): TestBlockHeaderValidation,
  TestGenesisBlock, TestPreValidation, TestReorgs, standalone tests,
  plus test_augmented_chain, test_get_block_generator, test_lookup_fork_chain

- blockchain.transactions (~1144 tests): TestBodyValidation (extracted
  to new file), test_blockchain_transactions, test_build_chains

Estimated critical path reduction: ~60 min -> ~40-45 min.
