+++ chia/_tests/blockchain/transactions/config.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+job_timeout = 80
+checkout_blocks_and_plots = True
